moved TPClient, movved contents of onListchange & OnConnectorChange to index
moved TPClient back to index.js
deleted onConnectorChange and onListChange and moved contents to index.js inside of TPClient
